Feat: Refactor the code. Move different logics into different folders. Organize the code
