+++ main.go
@@ -3,67 +3,40 @@ package main
 import (
 	"bufio"
 	"fmt"
-	"math/rand"
 	"os"
 	"strconv"
 	"strings"
+
+	"github.com/lielalmog/navy-six/menu"
+	"github.com/lielalmog/navy-six/missiles"
 )
 
 func cleanScreen() {
 	fmt.Print("\033[H\033[2J")
 }
 
-func convertStringToInt(s string) (int, error) {
-	s = strings.TrimSpace(s)
-
-	// Parse the string into an integer
-	number, err := strconv.Atoi(s)
-	if err != nil {
-		return 0, err
-	}
-	fmt.Println()
-
-	return number, nil
-}
-
-func readIntFromConsole(r *bufio.Reader) (int, error) {
-	input, err := r.ReadString('\n')
-	if err != nil {
-		return 0, err
-	}
-
-	// Trim the newline character from the input
-	input = strings.TrimSpace(input)
-
-	// Parse the string into an integer
-	number, err := strconv.Atoi(input)
-	if err != nil {
-		return 0, err
-	}
-	fmt.Println()
-
-	return number, nil
-}
-
-func random(min int, max int) int {
-	return rand.Intn(max-min) + min
-}
-
 func main() {
 	r := bufio.NewReader(os.Stdin)
-	initLaunchers()
+	missiles.InitLaunchers()
 
 	for {
 		cleanScreen()
-		printMenu()
+		menu.PrintMenu()
 
 		fmt.Print("Enter your choice: ")
-		choice, _ := readIntFromConsole(r)
-
-		if item, ok := menu[menuOption(choice)]; ok {
+		input, _ := r.ReadString('\n')
+		input = strings.TrimSpace(input)
+		number, err := strconv.Atoi(input)
+		if err != nil {
 			cleanScreen()
-			fmt.Printf("You selected: %s\n", item.name)
-			item.action()
+			fmt.Println("Invalid choice")
+			continue
+		}
+
+		if item, ok := menu.Menu[menu.MenuOption(number)]; ok {
+			cleanScreen()
+			fmt.Printf("You selected: %s\n", item.Name)
+			item.Action()
 
 			// Wait for user to press enter
 			fmt.Print("Press 'Enter' to continue...")
@@ -72,7 +45,7 @@ func main() {
 		} else {
 			cleanScreen()
 			fmt.Println("Invalid choice")
-			printMenu()
+			menu.PrintMenu()
 		}
 	}
 }
